+++ src/commonMain/kotlin/com/hoc081098/flowext/Event.kt
@@ -22,6 +22,8 @@
  * SOFTWARE.
  */
 
+@file:Suppress("NOTHING_TO_INLINE")
+
 package com.hoc081098.flowext
 
 /**
@@ -65,29 +67,52 @@ public inline fun <T, R> Event<T>.map(transform: (T) -> R): Event<R> = when (thi
 }
 
 /**
+ * Returns the result of applying [transform] to this [Event]'s [value][Event.Value.value] if this is a [Event.Value].
+ * Otherwise, returns itself.
+ *
+ * Slightly different from [map] in that [transform] is expected to
+ * return an [Event] (which could be [Event.Error] or [Event.Complete]).
+ */
+public inline fun <T, R> Event<T>.flatMap(transform: (T) -> Event<R>): Event<R> = when (this) {
+  Event.Complete -> Event.Complete
+  is Event.Error -> this
+  is Event.Value -> transform(value)
+}
+
+/**
  * Returns the encapsulated [value][Event.Value.value] if this [Event] is a [Event.Value], otherwise returns `null`.
  */
-public fun <T> Event<T>.valueOrNull(): T? = when (this) {
-  Event.Complete -> null
-  is Event.Error -> null
-  is Event.Value -> value
-}
+public inline fun <T> Event<T>.valueOrNull(): T? = valueOrElse { null }
+
+/**
+ * Returns the encapsulated [value][Event.Value.value] if this [Event] is a [Event.Value], otherwise returns [defaultValue].
+ */
+public inline fun <T> Event<T>.valueOrDefault(defaultValue: T): T? = valueOrElse { defaultValue }
 
 /**
  * Returns the encapsulated value if this [Event] is a [Event.Value].
  * If this is [Event.Error], throws the encapsulated [error][Event.Error.error].
  * Otherwise, throws a [NoSuchElementException].
  */
-public fun <T> Event<T>.valueOrThrow(): T = when (this) {
-  Event.Complete -> throw NoSuchElementException("$this has no value!")
-  is Event.Error -> throw error
+public inline fun <T> Event<T>.valueOrThrow(): T =
+  valueOrElse { throw it ?: NoSuchElementException("$this has no value!") }
+
+/**
+ * Returns the encapsulated [value][Event.Value.value] if this [Event] is a [Event.Value],
+ * otherwise returns the result of calling [defaultValue] function.
+ *
+ * The function [defaultValue] will be called with `null` if this [Event] is [Event.Complete].
+ */
+public inline fun <T> Event<T>.valueOrElse(defaultValue: (Throwable?) -> T): T = when (this) {
+  Event.Complete -> defaultValue(null)
+  is Event.Error -> defaultValue(error)
   is Event.Value -> value
 }
 
 /**
  * Returns the encapsulated [error][Event.Error.error] if this [Event] is an [Event.Error], otherwise returns `null`.
  */
-public fun <T> Event<T>.errorOrNull(): Throwable? = when (this) {
+public inline fun <T> Event<T>.errorOrNull(): Throwable? = when (this) {
   Event.Complete -> null
   is Event.Error -> error
   is Event.Value -> null
@@ -97,8 +122,5 @@ public fun <T> Event<T>.errorOrNull(): Throwable? = when (this) {
  * Returns the encapsulated [error][Event.Error.error] if this [Event] is an [Event.Error].
  * Otherwise, throws a [NoSuchElementException].
  */
-public fun <T> Event<T>.errorOrThrow(): Throwable = when (this) {
-  Event.Complete -> throw NoSuchElementException("$this has no error!")
-  is Event.Error -> error
-  is Event.Value -> throw NoSuchElementException("$this has no error!")
-}
+public inline fun <T> Event<T>.errorOrThrow(): Throwable =
+  errorOrNull() ?: throw NoSuchElementException("$this has no error!")